Fix compiler warnings about signed/unsigned.
git-svn-id: https://devel.aqbanking.de/svn/gwenhywfar/trunk@79 70169cfe-8b10-0410-8925-dcb4b91034d8

+++ src/base/buffer.c
@@ -606,7 +606,7 @@ int GWEN_Buffer_InsertBytes(GWEN_BUFFER *bf,
       return 0;
     }
     else {
-      if ((bf->ptr-bf->realPtr)>=size) {
+      if ( (bf->ptr - bf->realPtr) >= (int)size ) {
         /* we can simply insert it by occupying the reserved space */
         bf->ptr-=size;
         memmove(bf->ptr, buffer, size);
